Entregables/Leticia Quezada Calzada/APPIUM/requirements.txt
behave == 1.2.6
Appium-Python-Client==2.9.0
allure-behave==2.13.2
python-dotenv==1.0.0
Selenium==4.8.3
